+++ NEWS.md
@@ -4,6 +4,7 @@
 * Updated unit tests.
 * Added vignette on manually adjusting ABfit baseline smoothness.
 * Updated the introduction vignette.
+* Bug fixes for image and stackplot functions for masked MRS data.
 
 # spant 1.0.0
 * ABfit analysis method has been added, and is now default for the mrs_fit